requirements.txt
flask
flask-cors
python-dotenv
pandas
google-genai